ollama_GenAI/requirements.txt
python-dotenv

langchain
langchain_community

ipykernel

streamlit
watchdog

faiss-cpu

